+++ Bluetooth/BSN/include/bm_blemesh.h
@@ -1,0 +1,12 @@
+#ifndef BM_BLEMESH_H
+#define BM_BLEMESH_H
+
+
+
+/* Init and enable the BLE Mesh Stack 
+* 
+*	Does Self Provisioning
+*/
+void bm_blemesh_enable(void);
+
+#endif
