--- Bluetooth/BSN/include/const.h
@@ -1,65 +1,0 @@
-/** Defines for Self Provisioning **/
-static const u8_t net_key[16] = {
-	0x66,
-	0xef,
-	0x1e,
-	0x3a,
-	0x18,
-	0x0d,
-	0x7d,
-	0xec,
-	0x15,
-	0x45,
-	0xe6,
-	0xae,
-	0x77,
-	0x47,
-	0xb3,
-	0xfe,
-};
-static u8_t dev_key[16] = {
-	0x01,
-	0x23,
-	0x45,
-	0x67,
-	0x89,
-	0xab,
-	0xcd,
-	0xef,
-	0x01,
-	0x23,
-	0x45,
-	0x67,
-	0x89,
-	0xab,
-	0xcd,
-	0xef,
-};
-static const u8_t app_key[16] = {
-	0xa3,
-	0x01,
-	0x77,
-	0x4b,
-	0xe8,
-	0x02,
-	0x2c,
-	0xd8,
-	0x12,
-	0x3f,
-	0xf3,
-	0x27,
-	0x42,
-	0xa7,
-	0x36,
-	0x94,
-};
-
-static const u16_t net_idx=0;
-static const u16_t app_idx=0;
-static const u32_t iv_index=0;
-
-#define GROUP_ADDR 0xc000 //Range from 0xC000-0xFEFF 
-
-static u8_t flags=0;
-static u16_t addr = 0x0b0d; // Unicast Address Range from 0-0x7FFF, 0-32767 ->Assigned Random
-
