+++ Bluetooth/BSN/src/main.c
@@ -1,0 +1,7 @@
+#include <zephyr.h>
+#include "bm_blemesh.h"
+
+void main(void)
+{
+	bm_blemesh_enable();
+}
